chore: Refactor KhipuPay initialization method
The KhipuPay initialization method has been refactored to simplify the code and improve readability. The `KeyMode` parameter has been removed, and the `apiKey` parameter is now optional. If an `apiKey` is provided, it will be used to initialize the instance. Otherwise, the `apiKey` will be retrieved from the environment variables. If no `apiKey` is found, an error will be thrown.